pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gitdude"
version = "1.0.3"
description = "AI-powered Git workflow assistant — commit, review, recover, and more with a single command"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "GitDude Contributors" }]
keywords = ["git", "ai", "cli", "developer-tools", "gemini", "groq", "openai", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
requires-python = ">=3.10"

dependencies = [
    "typer[all]>=0.12.0",
    "gitpython>=3.1.40",
    "rich>=13.7.0",
    "pyperclip>=1.8.2",
    "google-generativeai>=0.7.0",
    "groq>=0.9.0",
    "ollama>=0.2.0",
    "openai>=1.30.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
gitdude = "gitdude.main:main"
gitwise = "gitdude.main:main"

[project.urls]
Homepage = "https://github.com/utkarshgupta188/gitdude"
Repository = "https://github.com/utkarshgupta188/gitdude"
Documentation = "https://github.com/utkarshgupta188/gitdude#readme"
"Bug Tracker" = "https://github.com/utkarshgupta188/gitdude/issues"

[tool.hatch.build.targets.wheel]
packages = ["gitdude"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
